Renamed student parameters

+++ src/main/java/com/example/demo/service/StudentService.java
@@ -7,9 +7,9 @@ import com.example.demo.entity.Student;
 public interface StudentService {
 
     public void addNewStudent(Student student);
-    public void deleteStudent(Long studentId);
-    public void updateStudent(Long studentId, String name, String email);
-    public Student getStudentById(Long studentId);
+    public void deleteStudent(Integer studentId);
+    public void updateStudent(Integer studentId, String name, String email);
+    public Student getStudentById(Integer studentId);
     public Student getStudentByEmail(String email);
     public List<Student> getStudents();
